Update GlassFalling.java
Added comment to show that I didn't do the "optional" Memoized, therefore I had to comment out the minTrials2Recur in the output so the recursion doesn't keep running for a long time. I've spoken to you after class regarding this and just wanted to let you know as to why its been commented out.